Fix: Construct the message within the service level
