+++ index.js
@@ -1,8 +1,7 @@
 require('dotenv').config();
 const express = require('express');
 const bodyParser = require('body-parser');
-const routes = require('./routes')
-
+const routes = require('./routes');
 
 const app = express();
